Cmake buildsystem migration (#41)
* Add initial CMake build scripts

Everything builds and links, assuming our Python SDK is found. Still need to source that correctly, which is WIP

* Add CCP Stackless Python

Everything builds now, and tests can be run. There are 6 test failures though that need to be looked into.

* Add platform specific guards to python configuration

* Update ccp-python.cmake to reflect updated framework build

* Run tests as post-build step

* Fix wrong windows size in test_summary_window

* Add missing HAVE_ZLIB compile option

* Add Windows support to our CCP Python package

* Add Windows builds

* Fix macOS paths and platform detection for ccp stackless python package

* Install pytest into user folder

Should fix the build failure we recently saw on the TeamCity macOS agent. The TeamCity macOS agents have an older version installed in the system environment because of matplotlib.

* Add explicit HAVE_ZLIB define back in

Apparently it's not exposed on the import target created by CMake's find_package module for ZLIB.

* Change to public include directory

Since we're including from it in a header file, so it should be visible to targets which use the `prometheusmodule` target.

* Use `ccpgames/platform-cmake` to source Python

* Use add_python_extension helper

* Fix module name

add_python_module uses the target name as output name for the module by default.

* Update vendored ccpgames/platform-cmake

So that we get the version of FindPython.cmake that contains our include guards.

* Add option to disable building tests

Useful when using this project from other CMake projects.

* Add support for building with CMake on Windows

This commit assumes that all required dependencies, except Python, are already available in the build environment, for example via `vcpkg`. This also enables us to clean up some of the CMake instructions which accessed private implementation details of dependencies sourced via the `FetchContent` module.

Note: 6 of the tests are currently failing for me on Windows, will need to look into it.

* Fix casing in project / target names

Co-authored-by: Thomas Dähling <24817+daehlith@users.noreply.github.com>