Add macOS support (#47)
* Update to latest platform-cmake module

So that we can use vendored modules from perforce, ensure system requirements, etc.

* Fix some issues highlighted by clang

A few `writable-strings` warnings, and some style issues, e.g. use `auto` and `nullptr`.

* Use Perforce vendored libraries for building

* Remove accidentally committed, obsolete patch for prometheus-cpp 0.7.0

* Fix out-of-bounds memory access in MetricsFactory

The `MetricFactory::Make*` functions could end up accessing an invalid iterator when determining the metric family, which (rightfully) bombs when running on macOS. With this change we ensure that the iterator is now valid before we try to access it.

* Fix running tests as post-build step

This broke when the file was accidentally committed in 1ae47223. So we restore that functionality and disable the cache provider plugin which tends to cause us issues when running through pythonInterpreter on Windows.

* Fix writable-strings warnings

* Enable verbose logging when apply git patch

For some reason this fails on the TeamCity Windows build, and I cannot reproduce this on my local. So let's get more output.

* Add some debug logging

Let's see what git configuration the job runs with, maybe there's an old-fashioned line ending issue preventing the patch from applying.

* Disable git's core.autocrlf when applying patch files

It seems that this is the main difference between local Windows workstation and the TeamCity agent. So let's see if disabling it here makes TeamCity apply the patch.

* Place test_results.xml in test folder

* Use CMAKE_CURRENT_*_DIR as opposed to CMAKE_*_DIR

CMAKE_*_DIR refers to the source and binary paths of the top level project. However, when sourcing a project via FetchContent, then this is not the top level project. Thus, we want to use the CMAKE_CURRENT_*_DIR variables which will always point at the location of the currently executing CMake script.

This also removes the debug code that was introduced for teamcity. As it turned out, disabling core.autocrlf did not change anything. Instead, the build failure we saw on Windows was due to using the Visual Studio generator.

* Use correct target name for CURL

This was wrong because the monolith's vendored curlConfig.cmake was wrong.

* Add project guard around test targets

So that tests are not built by default when including this project via FetchContent.

* Remove debug code

There was a spurious print statement left in.

* Update to latest prometheus-cpp release

Version 0.7.0 does not build out of the box due to its outdated civetweb dependency. Tests are currently broken after the update, going through them one by one.

* Remove "Gauge must not decrement" assertion

This limitation was removed in the latest version of prometheus-cpp to be more in line with the golang library. See https://github.com/jupp0r/prometheus-cpp/pull/406/

* Remove port format tests

They are testing Civetweb's implementation and are thus completely ouf of place for this module.

* Remove another Gauge assertion

Forgot to remove it in the previous commit, but it's the same reason.

* Set an explicit timeout when reading results in tests

The summary window test kept failing for me locally, and after instrumenting the code for it then it turned out that the `urllib2.urlopen()` call would take about a second each. This in turn caused the quantile to _just_ miss the 4 seconds window, failing the test.

* Update prometheus-cpp to v1.0.0

Since the v0.7.0 release didn't build for us on Windows for some unknown reason. The vcpkg version seems to have worked just fine, but the official release from GitHub failed due to its CivetWeb dependency failing to build.

This also highlighted some issues in the provided tests, particularly `test_counters_with_different_label_values_share_one_type_definition` failed. Not entirely sure why, but since it was inconsistent with the same test for the other metric types then I simply adapted the counter version to be working the same way.

* Remove unused functions

cppcheck was complaining about these being unused. Which is true, since the C++ tests for it were removed in 0f5c2c37d13c18580ef38066f172f935372cdffa

* Disable prometheus-cpp tests

We don't really need to run these ourselves, and they were causing our builds to fail because they found a different set of cmake packages. This also uncovered a problem with the way that the test_native target was linking its libraries, as it seems to have accidentally used the googletest package from prometheus-cpp as opposed to the one we included ourselves.

* Remove obsolete prometheus-cpp patch

This was required for version 0.7 of prometheus-cpp. The 1.0 release uses more modern cmake which makes this patch unnecessary.