Add DLL dependencies for tests to search paths
Tests were failing to run due to not being able to find the GTest DLLs. As of Python 3.8, adding DLLs to the PATH does not enable them to be picked up by imported modules. Rather than copying the DLLs around, populate an environment variable from CMake and use it to populate DLL search paths before running the tests.